# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toya Amerta API",
        "description": "REST API untuk aplikasi mobile (Flutter) Sistem Informasi PDAM Desa Toya Amerta. Semua response menggunakan envelope { success, message, data, meta }.",
        "contact": {
            "name": "Toya Amerta Dev Team",
            "email": "pt.andisuartika@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/petugas/maintenance": {
            "get": {
                "tags": [
                    "Petugas - Maintenance"
                ],
                "summary": "Daftar laporan maintenance",
                "operationId": "fcab5457fe4e9f1efbaa36712e62a4fb",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dilaporkan",
                                "dalam_proses",
                                "selesai",
                                "ditunda"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter prioritas",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "rendah",
                                "sedang",
                                "tinggi",
                                "darurat"
                            ]
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter kategori",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pipa_bocor",
                                "meteran_rusak",
                                "pompa",
                                "reservoir",
                                "instalasi_baru",
                                "lainnya"
                            ]
                        }
                    },
                    {
                        "name": "zone_id",
                        "in": "query",
                        "description": "Filter zona",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Maintenance"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Petugas - Maintenance"
                ],
                "summary": "Buat laporan maintenance",
                "operationId": "049b4b08cf092efe9df0b8a08caa91b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "category",
                                    "priority",
                                    "reported_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Pipa bocor di Jl. Raya Desa"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Jl. Raya Desa No. 45"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "pipa_bocor",
                                        "enum": [
                                            "pipa_bocor",
                                            "meteran_rusak",
                                            "pompa",
                                            "reservoir",
                                            "instalasi_baru",
                                            "lainnya"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "tinggi",
                                        "enum": [
                                            "rendah",
                                            "sedang",
                                            "tinggi",
                                            "darurat"
                                        ]
                                    },
                                    "reported_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-08"
                                    },
                                    "zone_id": {
                                        "type": "integer",
                                        "example": 2,
                                        "nullable": true
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Air merembes dari sambungan pipa bawah tanah.",
                                        "nullable": true
                                    },
                                    "material_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "labor_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Laporan maintenance berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Laporan maintenance berhasil disimpan."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Maintenance"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/petugas/maintenance/{id}": {
            "get": {
                "tags": [
                    "Petugas - Maintenance"
                ],
                "summary": "Detail laporan maintenance",
                "operationId": "389a26ad048b83ddf0b83797adee2aa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID laporan maintenance",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Maintenance"
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role tidak diizinkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedResponse": {
                "title": "Unauthenticated Response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated. Token tidak valid atau sudah kedaluwarsa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Maintenance": {
                "title": "Maintenance",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 8
                    },
                    "title": {
                        "type": "string",
                        "example": "Pipa bocor di Jl. Raya Desa"
                    },
                    "location": {
                        "type": "string",
                        "example": "Jl. Raya Desa No. 45"
                    },
                    "category": {
                        "type": "string",
                        "example": "pipa_bocor",
                        "enum": [
                            "pipa_bocor",
                            "meteran_rusak",
                            "pompa",
                            "reservoir",
                            "instalasi_baru",
                            "lainnya"
                        ]
                    },
                    "category_label": {
                        "type": "string",
                        "example": "Pipa Bocor"
                    },
                    "priority": {
                        "type": "string",
                        "example": "tinggi",
                        "enum": [
                            "rendah",
                            "sedang",
                            "tinggi",
                            "darurat"
                        ]
                    },
                    "priority_label": {
                        "type": "string",
                        "example": "Tinggi"
                    },
                    "status": {
                        "type": "string",
                        "example": "dilaporkan",
                        "enum": [
                            "dilaporkan",
                            "dalam_proses",
                            "selesai",
                            "ditunda"
                        ]
                    },
                    "status_label": {
                        "type": "string",
                        "example": "Dilaporkan"
                    },
                    "zone": {
                        "type": "string",
                        "example": "Zona B",
                        "nullable": true
                    },
                    "customer_name": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "reported_by": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "reported_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-08"
                    },
                    "handled_date": {
                        "type": "string",
                        "format": "date",
                        "example": null,
                        "nullable": true
                    },
                    "completed_date": {
                        "type": "string",
                        "format": "date",
                        "example": null,
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Air merembes dari sambungan pipa bawah tanah.",
                        "nullable": true
                    },
                    "material_cost": {
                        "type": "number",
                        "format": "float",
                        "example": null,
                        "nullable": true
                    },
                    "labor_cost": {
                        "type": "number",
                        "format": "float",
                        "example": null,
                        "nullable": true
                    },
                    "total_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token yang didapat dari endpoint /auth/login. Format header: Authorization: Bearer {token}",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}